refactor: calculate range of values to use based on indices

+++ pumpkin-crates/conflict-resolvers/src/minimisers/semantic_minimiser.rs
@@ -41,7 +41,7 @@ create_statistics_struct!(SemanticMinimiserStatistics {
 
 impl Default for SemanticMinimiser {
     fn default() -> Self {
-        let mapping = |x: &DomainId| x.id() as usize;
+        let mapping = |x: &DomainId| x.id() as i32;
         Self {
             original_domains: Default::default(),
             domains: Default::default(),
